chore(test): Move array and associative array tests into brsTypes/components
Now this directory tree matches the other one yaaaaay
